requirements.txt
openai>=1.0.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
